Added scriptError event to make it easier to find errors with scripts (scriptError(errorString, callstack (table with the function name as the index and an array with the script name and line as the value), locals (array with the local name and value)). Added example for scriptError event to cp.nut. Moved dimension id stuff to Common.h. Fixed VehicleEntryExit rpc stuff. Added vehicle SetDimension rpc to CVehicle::SetDimension. Removed CServer::OnUnload terminating the process. Made a few minor changes to CSquirrelArgument and CSquirrelArguments classes.

+++ trunk/Shared/CEvents.h
@@ -231,14 +231,14 @@ public:
 
 	CSquirrelArgument Call(String strName, CSquirrel* pScript = NULL)
 	{
-		CSquirrelArgument pReturn(1);
+		CSquirrelArgument pReturn((int)1);
 		Call(strName, &CSquirrelArguments(), &pReturn, pScript);
 		return pReturn;
 	}
 
 	CSquirrelArgument Call(String strName, CSquirrelArguments* pArguments, CSquirrel* pScript = NULL)
 	{
-		CSquirrelArgument pReturn(1);
+		CSquirrelArgument pReturn((int)1);
 		Call(strName, pArguments, &pReturn, pScript);
 		return pReturn;
 	}
